requirements.txt
streamlit>=1.28.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-core>=0.1.0
chromadb>=0.4.22
sentence-transformers>=2.2.2
pypdf>=3.17.0
python-dotenv>=1.0.0
unstructured>=0.10.30
langchain-text-splitters>=0.0.1
python-magic-bin>=0.4.14  # Windows only
python-magic>=0.4.27      # macOS/Linux
fastapi>=0.104.0
uvicorn>=0.24.0
requests
